Change in detailed report

+++ src/main/java/com/ClinicaDelCalzado_BackEnd/dtos/detailedReport/DetailedReportDTO.java
@@ -7,8 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Getter
 @Builder
 @NoArgsConstructor
@@ -18,9 +16,9 @@ public class DetailedReportDTO {
 
     private String creationDate;
     private String orderNumber;
-    private Double totalServicesValue;
-    private Double totalDeposits;
-    private Double totalBalance;
+    private Long totalServicesValue;
+    private Long totalDeposits;
+    private Long totalBalance;
     private Long servicesReceived;
     private Long servicesCompleted;
     private Long servicesDispatched;
